Update generate from prompt list embeddings, update readme, modularize generate from prompt list embeddings
